+++ docs/root/intro/version_history.rst
@@ -24,6 +24,7 @@ Version history
 * thrift_proxy: added support for cluster header based routing.
 * tls: remove TLS 1.0 and 1.1 from client defaults
 * router: exposed DOWNSTREAM_REMOTE_ADDRESS as custom HTTP request/response headers.
+* router: added support for percentage-based retry budgets
 
 1.12.0 (October 31, 2019)
 =========================